Commit for IGenericMetier

+++ src/main/java/com/inti/formation/IMetier/IGenericMetier.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface IGenericMetier<G> {
 	public void save(G g);
 	public void update(G g);
-	public void delete(Long id);
-	public G getOne(Long id);
+	public void delete(int id);
+	public G getOne(int id);
 	public List<G> getAll();
 }
